fix(catalog): the columns PG declares as "char", and the type itself
sentori's 5.1. 28 catalog columns read text where PG 18.6 declares
"char", so a client that types its result columns saw the wrong OID for
relkind, contype, typtype, castmethod and the rest. They carry the type
now, values included; the letters are unchanged.

Measuring it found the type barely reachable and three things wrong
with it: CREATE TABLE t(k "char") was a syntax error, the integer
conversions were unsigned (((-128)::int::"char")::int answered 128),
and charout printed the byte as a char -- so 0 became a NUL inside the
string and 0x80 became a two-byte character where PG prints \200.

And psql \d <table> failed outright: its constraint query ends with
VALUES ('16384'::pg_catalog.regclass) and SPG looked that up as a
relation name. PG's regclassin reads an all-digit string as the OID and
the object need not exist; so does SPG's now, and regtypein with it. \d
on a table with a primary key and an index now prints byte-identically
to PG 18.6's.

5 pins; four ablations bite.

+++ crates/spg-engine/src/eval/functions.rs
@@ -3484,7 +3484,8 @@ fn apply_function_dispatch(
                     Ok(Value::Int(s.chars().next().map_or(0, |ch| ch as i32)))
                 }
                 // v7.39 (read01 char.c) — ascii("char") is the byte value.
-                Value::Char1(b) => Ok(Value::Int(i32::from(*b))),
+                // 9.0.0 — signed; see the cast arm.
+                Value::Char1(b) => Ok(Value::Int(i32::from(*b as i8))),
                 other => Err(EvalError::TypeMismatch {
                     detail: alloc::format!(
                         "ascii() needs text, got {}",
